+++ src/main/java/ca/tyny/urlshortener/infra/security/AuthenticationAdapter.java
@@ -1,0 +1,22 @@
+package ca.tyny.urlshortener.infra.security;
+
+import ca.tyny.urlshortener.core.ports.outgoing.AuthenticationPort;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.stereotype.Component;
+
+@Component
+public class AuthenticationAdapter implements AuthenticationPort {
+
+    private final AuthenticationManager authenticationManager;
+
+    public AuthenticationAdapter(AuthenticationManager authenticationManager) {
+        this.authenticationManager = authenticationManager;
+    }
+
+    @Override
+    public void authenticate(String email, String password) {
+        authenticationManager.authenticate(
+                new UsernamePasswordAuthenticationToken(email, password));
+    }
+}